:bug: fix: Divide payment transaction with facade pattern
