+++ src/main/java/nbc/mushroom/domain/payment/controller/PaymentController.java
@@ -7,7 +7,7 @@ import nbc.mushroom.domain.common.dto.ApiResponse;
 import nbc.mushroom.domain.common.dto.AuthUser;
 import nbc.mushroom.domain.payment.dto.request.PaymentReq;
 import nbc.mushroom.domain.payment.dto.response.PaymentRes;
-import nbc.mushroom.domain.payment.service.PaymentService;
+import nbc.mushroom.domain.payment.service.PaymentFacade;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -20,14 +20,14 @@ import org.springframework.web.bind.annotation.RestController;
 @RequestMapping("/api/payments")
 public class PaymentController {
 
-    private final PaymentService paymentService;
+    private final PaymentFacade paymentFacade;
 
     @PostMapping("/confirm")
     public ResponseEntity<ApiResponse<PaymentRes>> confirmPayment(
         @Auth AuthUser authUser,
         @Valid @RequestBody PaymentReq paymentReq
     ) {
-        PaymentRes paymentRes = paymentService.confirmPayment(authUser, paymentReq);
+        PaymentRes paymentRes = paymentFacade.confirmPayment(authUser, paymentReq);
 
         return ResponseEntity.status(HttpStatus.OK)
             .body(ApiResponse.success("결제가 정상적으로 완료되었습니다.", paymentRes));
